refactor each node type to individual files
