+++ src/parser/ast_node/ast_node_literal.rs
@@ -1,0 +1,10 @@
+#[derive(Debug)]
+pub(crate) struct ASTNodeLiteral {
+    m_value: char,
+}
+
+impl PartialEq for ASTNodeLiteral {
+    fn eq(&self, other: &Self) -> bool {
+        self.m_value == other.m_value
+    }
+}
